+++ HISTORY.rst
@@ -16,6 +16,7 @@ Historia zmian
 * poprawki pobierania PubMed ID (#958),
 * poprawne zawężanie do zakresu punktów PK (#967),
 * katalog cache ma nazwę z numerem wersji (#961),
+* raport slotów uczelnia wg algorytmu plecakowego (#923),
 * w multiwyszukiwarce w polu "Wydawnictwo nadrzędne" pokazuj wyłącznie rekordy
   będące już wydawnictwami nadrzędnymi dla rekordów (#953).